feat: implement analyzeImagesWithOkHttp for analyzing images via API
