+++ app/src/main/java/com/android/solvit/seeker/ui/request/RequestComposables.kt
@@ -1,6 +1,5 @@
 package com.android.solvit.seeker.ui.request
 
-import analyzeImagesWithOpenAI
 import android.net.Uri
 import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -519,323 +518,4 @@ fun DeleteButton(
               Text("Delete", color = colorScheme.error)
             }
       }
-}
-
-@Composable
-fun AIAssistantDialog(onCancel: () -> Unit, onUploadPictures: () -> Unit) {
-  Dialog(onDismissRequest = { onCancel() }) {
-    Box(
-        modifier =
-            Modifier.fillMaxWidth(0.9f) // Adjust width as 90% of the screen
-                .wrapContentHeight()
-                .background(Color.White, shape = RoundedCornerShape(16.dp))
-                .padding(16.dp) // Padding around the dialog content
-        ) {
-          Column(
-              modifier = Modifier.fillMaxWidth(),
-              horizontalAlignment = Alignment.CenterHorizontally) {
-                // Header Row
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.padding(bottom = 4.dp)) {
-                      // Icon
-                      Box(
-                          modifier =
-                              Modifier.size(64.dp)
-                                  .background(
-                                      Color(0xFF2196F3), shape = CircleShape), // Blue Circle
-                          contentAlignment = Alignment.Center) {
-                            Icon(
-                                painter =
-                                    painterResource(
-                                        id =
-                                            R.drawable.ic_ai_assistant), // Replace with proper icon
-                                contentDescription = "AI Assistant",
-                                tint = Color.White,
-                                modifier = Modifier.size(32.dp))
-                          }
-
-                      // Title
-                      Box(
-                          modifier =
-                              Modifier.offset(x = -14.dp)
-                                  .fillMaxWidth()
-                                  .background(
-                                      Color(0xFF2196F3),
-                                      shape =
-                                          RoundedCornerShape(
-                                              topStart = 0.dp,
-                                              topEnd = 18.dp,
-                                              bottomStart = 0.dp,
-                                              bottomEnd = 18.dp))
-                                  .padding(horizontal = 8.dp, vertical = 10.dp)) {
-                            Text(
-                                text = "Your AI-Powered Assistant",
-                                style =
-                                    MaterialTheme.typography.bodyMedium.copy(
-                                        fontWeight = FontWeight.Bold),
-                                color = Color.White,
-                                textAlign = TextAlign.Center)
-                          }
-                    }
-
-                // Dialog Content
-                Text(
-                    text =
-                        "Would you like to use the AI assistant to create your request by uploading pictures of your issue?",
-                    style = MaterialTheme.typography.bodyMedium,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier.padding(vertical = 16.dp))
-
-                // Buttons
-                Row(
-                    horizontalArrangement = Arrangement.Center,
-                    modifier = Modifier.fillMaxWidth()) {
-                      Button(
-                          onClick = { onCancel() },
-                          colors = ButtonDefaults.buttonColors(containerColor = Color.Red),
-                          modifier = Modifier.padding(end = 8.dp) // Space between buttons
-                          ) {
-                            Text("Cancel")
-                          }
-
-                      Button(
-                          onClick = { onUploadPictures() },
-                          colors = ButtonDefaults.buttonColors(containerColor = Color.Green)) {
-                            Text("Upload Pictures")
-                            Spacer(modifier = Modifier.width(4.dp))
-                            Icon(Icons.Default.ArrowForward, contentDescription = null)
-                          }
-                    }
-              }
-        }
-  }
-}
-
-@Composable
-fun ImagePickerStep(
-    selectedImages: List<Uri>,
-    onImagesSelected: (List<Uri>) -> Unit,
-    onRemoveImage: (Uri) -> Unit,
-    onStartAnalyzing: () -> Unit
-) {
-  val imagePickerLauncher =
-      rememberLauncherForActivityResult(ActivityResultContracts.GetMultipleContents()) { uris ->
-        if (uris != null) {
-          onImagesSelected(selectedImages + uris)
-        }
-      }
-
-  Column(modifier = Modifier.fillMaxWidth().padding(16.dp).verticalScroll(rememberScrollState())) {
-    StepHeader(R.drawable.circle_one_icon, title = "Upload Images")
-
-    Spacer(modifier = Modifier.height(16.dp)) // Spacing below the header
-
-    if (selectedImages.isEmpty()) {
-      Text(
-          text = "No Images added",
-          style = MaterialTheme.typography.bodyMedium,
-          textAlign = TextAlign.Center,
-          modifier = Modifier.fillMaxWidth())
-    } else {
-      LazyRow(
-          horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.fillMaxWidth()) {
-            items(selectedImages) { uri ->
-              Box(
-                  modifier =
-                      Modifier.size(80.dp)
-                          .clip(RoundedCornerShape(8.dp))
-                          .background(Color.Gray)
-                          .padding(4.dp)) {
-                    AsyncImage(
-                        model = uri, contentDescription = null, modifier = Modifier.fillMaxSize())
-                    Box(
-                        Modifier.align(Alignment.TopEnd)
-                            .background(Color.White, shape = CircleShape)
-                            .clickable { onRemoveImage(uri) }
-                            .padding(4.dp)) {
-                          Icon(
-                              Icons.Default.Delete,
-                              contentDescription = "Remove Image",
-                              tint = Color.Red)
-                        }
-                  }
-            }
-          }
-    }
-
-    Spacer(modifier = Modifier.height(16.dp)) // Spacing before the buttons
-
-    // Buttons Row
-    Row(horizontalArrangement = Arrangement.Center, modifier = Modifier.fillMaxWidth()) {
-      Button(
-          onClick = { imagePickerLauncher.launch("image/*") },
-          modifier = Modifier.weight(1f) // Ensures centering in the row
-          ) {
-            Text("Add Images")
-            Spacer(modifier = Modifier.width(4.dp))
-            Icon(Icons.Default.Add, contentDescription = null)
-          }
-
-      if (selectedImages.isNotEmpty()) {
-        Spacer(modifier = Modifier.width(8.dp)) // Space between buttons
-
-        Button(
-            onClick = { onStartAnalyzing() },
-            modifier = Modifier.weight(1f),
-            colors = ButtonDefaults.buttonColors(containerColor = Color.Green)) {
-              Text("Analyze")
-              Icon(Icons.Default.ArrowForward, contentDescription = null)
-            }
-      }
-    }
-  }
-}
-
-@Composable
-fun StepHeader(@DrawableRes iconRes: Int, title: String) {
-  // Header Row
-  Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.padding(bottom = 4.dp)) {
-    // Icon Number
-    Box(
-        modifier =
-            Modifier.size(60.dp).background(Color(0xFF2196F3), shape = CircleShape), // Blue Circle
-        contentAlignment = Alignment.Center) {
-          Icon(
-              painter = painterResource(id = iconRes), // Replace with proper icon
-              contentDescription = "Stepper Icon",
-              tint = Color.White,
-              modifier = Modifier.size(40.dp))
-        }
-
-    // Title
-    Box(
-        modifier =
-            Modifier.offset(x = -14.dp)
-                .fillMaxWidth()
-                .background(
-                    Color(0xFF2196F3),
-                    shape =
-                        RoundedCornerShape(
-                            topStart = 0.dp, topEnd = 18.dp, bottomStart = 0.dp, bottomEnd = 18.dp))
-                .padding(horizontal = 8.dp, vertical = 10.dp),
-        contentAlignment = Alignment.Center) {
-          Text(
-              text = title,
-              style = MaterialTheme.typography.bodyMedium.copy(fontWeight = FontWeight.Bold),
-              color = Color.White,
-              textAlign = TextAlign.Center)
-        }
-  }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewAIAssistantDialog() {
-  StepHeader(R.drawable.circle_one_icon, title = "Upload Images")
-}
-
-@Composable
-fun MultiStepDialog(
-    requestViewModel: ServiceRequestViewModel,
-    showDialog: Boolean,
-    currentStep: Int,
-    selectedImages: List<Uri>,
-    onImagesSelected: (List<Uri>) -> Unit,
-    onRemoveImage: (Uri) -> Unit,
-    onStartAnalyzing: () -> Unit,
-    onAnalyzeComplete: (String, String, String) -> Unit,
-    onClose: () -> Unit
-) {
-  var isLoading by remember { mutableStateOf(false) }
-  var showSparkleEffect by remember { mutableStateOf(false) } // Control for the sparkle effect
-
-  if (showDialog) {
-    Dialog(onDismissRequest = { onClose() }) {
-      Surface(shape = RoundedCornerShape(16.dp), modifier = Modifier.padding(16.dp)) {
-        when (currentStep) {
-          1 ->
-              ImagePickerStep(
-                  selectedImages = selectedImages,
-                  onImagesSelected = onImagesSelected,
-                  onRemoveImage = onRemoveImage,
-                  onStartAnalyzing = onStartAnalyzing)
-          2 -> {
-            LaunchedEffect(selectedImages) {
-              if (selectedImages.isNotEmpty() && !isLoading) {
-                isLoading = true
-                try {
-                  val (type, title, description) =
-                      uploadAndAnalyze(requestViewModel, selectedImages)
-                  onAnalyzeComplete(type, title, description)
-                } catch (e: Exception) {
-                  Log.e("MultiStepDialog", "Error: ${e.message}")
-                } finally {
-                  isLoading = false
-                }
-              }
-            }
-
-            if (isLoading) {
-              Column(
-                  modifier = Modifier.fillMaxWidth().padding(16.dp),
-                  horizontalAlignment = Alignment.CenterHorizontally) {
-                    StepHeader(R.drawable.circle_two_icon, title = "Analyzing Images")
-                    Spacer(modifier = Modifier.height(16.dp))
-                    CircularProgressIndicator(modifier = Modifier.size(48.dp))
-                    Spacer(modifier = Modifier.height(16.dp))
-                    Text("Analyzing your uploaded images. Please wait...")
-                  }
-            }
-          }
-          3 -> {
-            Column(
-                modifier =
-                    Modifier.fillMaxWidth().padding(16.dp).verticalScroll(rememberScrollState()),
-                horizontalAlignment = Alignment.CenterHorizontally) {
-                  StepHeader(R.drawable.circle_three_icon, title = "Analysis Completed")
-                  Spacer(modifier = Modifier.height(16.dp))
-                  Icon(
-                      imageVector = Icons.Default.CheckCircle,
-                      contentDescription = "Complete",
-                      tint = Color.Green,
-                      modifier = Modifier.size(48.dp))
-                  Spacer(modifier = Modifier.height(16.dp))
-                  Text(
-                      text = "Analysis complete! Fields have been filled with AI suggestions.",
-                      style = MaterialTheme.typography.bodyMedium,
-                      textAlign = TextAlign.Center)
-                }
-          }
-        }
-      }
-    }
-  }
-}
-
-suspend fun uploadAndAnalyze(
-    requestViewModel: ServiceRequestViewModel,
-    imageUris: List<Uri>
-): Triple<String, String, String> {
-  return withContext(Dispatchers.IO) {
-    try {
-      // Step 1: Upload images and get their URLs
-      val imageUrls =
-          suspendCoroutine<List<String>> { continuation ->
-            requestViewModel.uploadMultipleImages(
-                imageUris = imageUris,
-                onSuccess = { urls -> continuation.resume(urls) },
-                onFailure = { exception -> continuation.resumeWithException(exception) })
-          }
-
-      // Log uploaded image URLs
-      Log.i("uploadAndAnalyze", "Uploaded image URLs: $imageUrls")
-
-      // Step 2: Analyze uploaded images with OpenAI
-      analyzeImagesWithOpenAI(imageUrls)
-    } catch (e: Exception) {
-      Log.e("uploadAndAnalyze", "Error: ${e.message}", e)
-      throw Exception("Error during upload and analysis: ${e.message}")
-    }
-  }
 }
